requirements.txt
Scrapy
Django
Celery
mysqlclient
redis
django-rest-framework
django-filter
python-dotenv